Change the order of text input type and value attributes
This works around a rather obscure bug in Microsoft Edge.

Basically, I want to use an input with type date (knowing well that
not every browser implements it), so I use:

```elm
Element.Input.text [ htmlAttribute <| Html.Attributes.type_ "date" ]
```

This works well in most browsers, but it seems to trigger a weird edge
case bug in Microsoft Edge where it renders the date text _twice_, first
in ISO8601, then in the normal way it renders dates in inputs with type
date. So the text box contains both the texts "2019-03-08" and
"3/8/2019".

After quite a lot of experimentation, I found out that switching the
order of these two attributes fixes the problem. I can only speculate as
to why, but whatever it is it seems like it should be benign change to
make.